hackathon-template/pyproject.toml
[project]
name = "adk-made-simple"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google-adk==1.0.0",
]
